Added methods for admin filter form
Now the methods called by AdminTable are "formFilter" and "cleanFilter",
which by default call the "old" "formElement" and "clean". In this
way it is more easy to customize filters behavior.
It was possible to add a file filter which uses a text input and clean
the user submitted value as text in filters